halfway port to quicksilver
- the only tile that renders correctly is the loose tile
- none of the UI works so u have to launch the demo
- text rendering looks complicated so i'm not gonna for now

+++ src/tile.rs
@@ -84,8 +84,8 @@ impl ops::MulAssign for Direction {
     }
 }
 
-impl ops::Mul<Direction> for [f64; 2] {
-    type Output = [f64; 2];
+impl ops::Mul<Direction> for [f32; 2] {
+    type Output = [f32; 2];
 
     fn mul(self, rhs: Direction) -> Self::Output {
         let [x, y] = self;
